Public Documentation

Kernels Loading and Inspection Functions

Ephemerides.EphemerisProvider — Type

EphemerisProvider(file::String)
EphemerisProvider(files::Vector{String})

Create an EphemerisProvider instance by loading a single or multiple binary ephemeris kernel files specified by files. Currently, only NAIF Double precision Array File (DAF) kernels (i.e., SPK and PCK) are accepted.

Example

julia> eph = EphemerisProvider("PATH_TO_KERNEL")
EphemerisProvider([...])

julia> eph = EphemerisProvider(["PATH_TO_KERNEL_1", "PATH_TO_KERNEL_2"])
EphemerisProvider([])

Ephemerides.EphemRecordSPK — Type

EphemRecordSPK <: AbstractEphemRecord

Store the SPK metadata relative to a given (target, center) objects pair.

Fields

target – Int target object NAIF ID

center – Int center object NAIF ID

axes – Int reference axes ID

t_start – start times of each sub-window, in TDB seconds since J2000

t_end – final times of each sub-window, in TDB seconds since J2000

Ephemerides.EphemRecordPCK — Type

EphemRecordPCK <: AbstractEphemRecord

Store the PCK metadata relative to a given (target, center) axes pair.

Fields

target – Int target axes NAIF ID

center – Int center axes NAIF ID

t_start – start times of each sub-window, in TDB seconds since J2000

t_end – final times of each sub-window, in TDB seconds since J2000

Ephemerides.ephem_timescale_id — Function

ephem_timescale_id(eph::EphemerisProvider)

Retrieve a timescale ID associated with the ephemeris handler eph. It returns 1 for Barycentric Dynamical Time (TDB) and 2 for Barycentric Coordinate Time (TCB).

Warning

Ephemeris providers with mixed timescales are not supported. An error is thrown if in the ephemeris handler some segments are defined in TDB and some other segments in TCB.

Ephemerides.ephem_spk_timespan — Function

ephem_spk_timespan(eph::EphemerisProvider)

Return the minimum and maximum time available in the SPK kernels loaded within eph, in TDB seconds since J2000, together with a continuity parameter defined as follows:

0 no SPK data is available.

1 the quantities of all bodies are available for any time between the first and last time.

2 the quantities of some bodies are available on discontinuous time intervals between the first and last time.

3 the quantities of each body are available on a continuous time interval between the first and the last time, but not available for any time between the first and last time.

References

CALCEPH C++ library

See Also

See also ephem_pck_timespan.

Ephemerides.ephem_pck_timespan — Function

ephem_pck_timespan(eph::EphemerisProvider)

Return the minimum and maximum time available in the PCK kernels loaded within eph, in TDB seconds since J2000, together with a continuity parameter defined as follows:

0 no PCK data is available.

1 the quantities of all axes are available for any time between the first and last time.

2 the quantities of some axes are available on discontinuous time intervals between the first and last time.

3 the quantities of each axis are available on a continuous time interval between the first and the last time, but not available for any time between the first and last time.

References

CALCEPH C++ library

See Also

See also ephem_spk_timespan.

Ephemerides.ephem_spk_records — Function

ephem_spk_records(eph::EphemerisProvider)

Return a list of EphemRecordSPK storing metadata relative to each (target, center) object pairs in the loaded SPK kernels.

Ephemerides.ephem_pck_records — Function

ephem_pck_records(eph::EphemerisProvider)

Return a list of EphemRecordPCK storing metadata relative to each (target, center) axes pairs in the loaded PCK kernels.

Ephemerides.ephem_get_points — Function

ephem_get_points(eph::EphemerisProvider)

Return a list of NAIFIds representing bodies with available ephemeris data.

Ephemerides.ephem_get_axes — Function

ephem_get_axes(eph::EphemerisProvider)

Return a list of Frame IDs representing axes with available orientation data.

Computing Positions

Ephemerides.ephem_vector3 — Function

ephem_vector3(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 3-elements state of one body (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Ephemerides.ephem_vector6 — Function

ephem_vector6(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 6-elements state of one body (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Ephemerides.ephem_vector9 — Function

ephem_vector9(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 9-elements state of one body (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Ephemerides.ephem_vector12 — Function

ephem_vector12(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 12-elements state of one body (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Computing Orientation

Ephemerides.ephem_rotation3 — Function

ephem_rotation3(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 3-elements orientation angles of one set of axes (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Note

For the orientation angles, it is not possible to automatically compute the reverse transformation , i.e., if the orientation of PA440 is defined with respect to the ICRF, it is not possible to compute the rotation from the PA440 to the ICRF with this routine.

Ephemerides.ephem_rotation6 — Function

ephem_rotation6(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 6-elements orientation angles of one set of axes (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Note

For the orientation angles, it is not possible to automatically compute the reverse transformation , i.e., if the orientation of PA440 is defined with respect to the ICRF, it is not possible to compute the rotation from the PA440 to the ICRF with this routine.

Ephemerides.ephem_rotation9 — Function

ephem_rotation9(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 9-elements orientation angles of one set of axes (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Note

For the orientation angles, it is not possible to automatically compute the reverse transformation , i.e., if the orientation of PA440 is defined with respect to the ICRF, it is not possible to compute the rotation from the PA440 to the ICRF with this routine.

Ephemerides.ephem_rotation12 — Function

ephem_rotation12(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 12-elements orientation angles of one set of axes (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Note

For the orientation angles, it is not possible to automatically compute the reverse transformation , i.e., if the orientation of PA440 is defined with respect to the ICRF, it is not possible to compute the rotation from the PA440 to the ICRF with this routine.