Ephemeris

JSMDInterfaces.Ephemeris.AbstractEphemerisProvider — Type

AbstractEphemerisProvider

Abstract type to represent ephemeris types.

JSMDInterfaces.Ephemeris.EphemAxesRecord — Type

EphemAxesRecord

Store an orientation record metadata.

Fields

target – NAIF ID of the target axes.

jd_start – Julian date of the first time.

jd_end – Julian date of the last time.

axes – NAIF ID of the reference axes.

JSMDInterfaces.Ephemeris.EphemPointRecord — Type

EphemPointRecord

Store a position record metadata.

Fields

target – NAIF ID of the target point.

center – NAIF ID of the center point.

jd_start – Julian date of the first time.

jd_end – Julian date of the last time.

axes – NAIF ID of the reference axes.

JSMDInterfaces.Ephemeris.EphemerisError — Type

JSMDInterfaces.Ephemeris.EphemerisError

A type representing ephemeris load or computations errors.

JSMDInterfaces.Ephemeris.ephem_compute! — Method

ephem_compute!(res, eph::AbstractEphemerisProvider, jd0::Number, time::Number, 
    target::Int, center::Int, order::Int)

Abstract method to compute position and derivatives up to order of target with respect to center at the Julian Date jd0 + time.

Inputs

eph – Ephemeris provider.

jd0, time – jd0 + time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris.

target – The body or reference point whose coordinates are required.

center – The origin of the coordinate system.

order – The order of derivatives from 0 (position) to 3 (position, velocity, acceleration and jerk).

JSMDInterfaces.Ephemeris.ephem_orient! — Method

ephem_orient!(res, eph::AbstractEphemerisProvider, jd0::Number, time::Number, 
    target::Int, center::Int, order::Int)

Abstract method to compute Euler angles and derivatives up to order for the orientation of the target axes at epoch jd0 + time.

Inputs

eph – Ephemeris provider.

jd0, time – jd0 + time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris.

target – The axes whose orientation is required.

center – The parent set of axes.

order – The order of derivatives from 0 (angles) to 3 (angles, angles rate, etc...).

JSMDInterfaces.FilesIO.load — Method

load(::Type{<:AbstractEphemerisProvider}, files)

Load ephemeris files.