Frames

JSMDInterfaces.Frames.AbstractJSMDFrameGraph — Type

AbstractJSMDFrameGraph

Abstract type for frames graphs. Subtype it to create a new frames graph compatible with the ecosystem.

JSMDInterfaces.Frames.rotation12 — Method

rotation12(model::F, from::Int, to::Int, time::Number)

This function serves as an interface for constructing a rotation matrix, its first, second and third derivative between two axes within a frame graph model.

Arguments

model::F: An instance of a subtype of AbstractJSMDFrameGraph, representing the frame graph model.

from::Int: The source node or axis for the rotation matrix.

to::Int: The destination node or axis for the rotation matrix.

time::T: The independent variable (time), in seconds since J2000.

Warning

Concrete implementations of AbstractJSMDFrameGraph must provide this function!

JSMDInterfaces.Frames.rotation3 — Method

rotation3(model::F, from::Int, to::Int, time::Number)

This function serves as an interface for constructing a rotation matrix between two axes within a frame graph model.

Arguments

model::F: An instance of a subtype of AbstractJSMDFrameGraph, representing the frame graph model.

from::Int: The source node or axis for the rotation matrix.

to::Int: The destination node or axis for the rotation matrix.

time::T: The independent variable (time), in seconds since J2000.

Warning

Concrete implementations of AbstractJSMDFrameGraph must provide this function!

JSMDInterfaces.Frames.rotation6 — Method

rotation6(model::F, from::Int, to::Int, time::Number)

This function serves as an interface for constructing a rotation matrix and its first derivative between two axes within a frame graph model.

Arguments

model::F: An instance of a subtype of AbstractJSMDFrameGraph, representing the frame graph model.

from::Int: The source node or axis for the rotation matrix.

to::Int: The destination node or axis for the rotation matrix.

time::T: The independent variable (time), in seconds since J2000.

Warning

Concrete implementations of AbstractJSMDFrameGraph must provide this function!

JSMDInterfaces.Frames.rotation9 — Method

rotation9(model::F, from::Int, to::Int, time::Number)

This function serves as an interface for constructing a rotation matrix, its first, second derivative between two axes within a frame graph model.

Arguments

model::F: An instance of a subtype of AbstractJSMDFrameGraph, representing the frame graph model.

from::Int: The source node or axis for the rotation matrix.

to::Int: The destination node or axis for the rotation matrix.

time::T: The independent variable (time), in seconds since J2000.

Warning

Concrete implementations of AbstractJSMDFrameGraph must provide this function!

JSMDInterfaces.Frames.vector12 — Method

vector12(model::F, from::Int, to::Int, axis::Int, time::Number)

This function serves as an interface for constructing a position, velocity, acceleration, and jerk vector using a frame graph model.

Arguments

model::F: An instance of a subtype of AbstractJSMDFrameGraph, representing the frame graph model.

from::Int: The source node or point for the vector.

to::Int: The destination node or point for the vector.

axis::Int: The specific axis of the vector.

time::T: The independent variable (time), in seconds since J2000.

Warning

Concrete implementations of AbstractJSMDFrameGraph must provide this function!

JSMDInterfaces.Frames.vector3 — Method

vector3(model::F, from::Int, to::Int, axis::Int, time::Number)

This function serves as an interface for constructing a position vector using a frame graph model.

Arguments

model::F: An instance of a subtype of AbstractJSMDFrameGraph, representing the frame graph model.

from::Int: The source node or point for the vector.

to::Int: The destination node or point for the vector.

axis::Int: The specific axis of the vector.

time::T: The independent variable (time), in seconds since J2000.

Warning

Concrete implementations of AbstractJSMDFrameGraph must provide this function!

JSMDInterfaces.Frames.vector6 — Method

vector6(model::F, from::Int, to::Int, axis::Int, time::Number)

This function serves as an interface for constructing a position and velocity vector using a frame graph model.

Arguments

model::F: An instance of a subtype of AbstractJSMDFrameGraph, representing the frame graph model.

from::Int: The source node or point for the vector.

to::Int: The destination node or point for the vector.

axis::Int: The specific axis of the vector.

time::T: The independent variable (time), in seconds since J2000.

Warning

Concrete implementations of AbstractJSMDFrameGraph must provide this function!

JSMDInterfaces.Frames.vector9 — Method

vector9(model::F, from::Int, to::Int, axis::Int, time::Number)

This function serves as an interface for constructing a position, velocity, and acceleration vector using a frame graph model.

Arguments

model::F: An instance of a subtype of AbstractJSMDFrameGraph, representing the frame graph model.

from::Int: The source node or point for the vector.

to::Int: The destination node or point for the vector.

axis::Int: The specific axis of the vector.

time::T: The independent variable (time), in seconds since J2000.

Warning

Concrete implementations of AbstractJSMDFrameGraph must provide this function!