Models

JSMDInterfaces.Models.AbstractJSMDModel — Type

AbstractJSMDModel

Basic abstract type for every JSMD compatible model.

JSMDInterfaces.Models.AbstractJSMDModelData — Type

AbstractJSMDModelData

Basic abstract type for every JSMD compatible model data.

JSMD ModelData types are the interface of the JSMD environment with the external world, and are meant to provide a model agnostic interface to any customly formatted file.

The main objective of this data type is to create a JSMD simulation compatible Model by means of the parse_model interface.

JSMDInterfaces.Models.dump_model — Method

dump_model(::AbstractJSMDModel, ::N, args...; 
    kargs...) where {N <: AbstractArchiveNode}

This function serves as an interface to dump to an archive node a JSMD compatible model.

JSMDInterfaces.Models.load_model — Method

load_model(::Type{T}, ::Type{ <: AbstractJSMDModel}, ::N, 
    args...; kargs...) where {T, N <: AbstractArchiveNode}

This function serves as an interface to load from an archive node a JSMD compatible model.

JSMDInterfaces.Models.parse_data — Method

parse_data(::Type{T}, ::Type{D}, ::AbstractString; 
    kargs...) where {T, D <: AbstractJSMDModelData} end

This function serves as an interface to create a JSMD compatible model data object from a file.

JSMDInterfaces.Models.parse_model — Method

parse_model(::Type{T}, ::Type{M}, ::Type{D}, args...; 
    kargs...) where {T, D <: AbstractJSMDModelData, M <: AbstractJSMDModel}

This function serves as an interface to create a JSMD compatible model type from a ModelData type.