State Transformations

Classical Orbital Elements (coe)

AstroRepresentations.convert6_cart_to_coe — Function

convert6_cart_to_coe(sv::AbstractVector{<:Number}, μ::Number, [args]...)

Convert cartesian state representation into classical orbital elements.

Inputs

sv – state vector – L, T

μ – center gravitational parameter – L³/T²

Output

Classical orbital elements representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications. Vol. 12. Springer Science & Business Media, 2001.

AstroRepresentations.convert6_coe_to_cart — Function

convert6_coe_to_cart(sv::AbstractVector{<:Number}, μ::Number, [args]...)

Convert classical orbital elements state vector to cartesian state.

Inputs

sv – Keplerian elements – L, rad

μ – Center's gravitational parameter – L³/T²

Output

Cartesian representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications. Vol. 12. Springer Science & Business Media, 2001.

Equinoctial Orbital Elements (equi)

AstroRepresentations.convert6_cart_to_equi — Function

convert6_cart_to_equi(sv::AbstractVector{<:Number}, μ::Number, [args]...)

Convert cartesian state vector to Equinoctial keplerian elements.

Inputs

sv – state vector – L, T

μ – center gravitational parameter – L³/T²

Output

Equinoctial representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.

Cefola - DOI: 10.2514/6.1972-937

AstroRepresentations.convert6_equi_to_cart — Function

convert6_equi_to_cart(equi::AbstractVector{<:Number}, μ::Number, [args]...)

Convert equinoctial state elements to cartesian state.

Inputs

equi – Equinoctial elements – L, rad

μ – center gravitational parameter – L³/T²

Output

Cartesian representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.

Cefola - DOI: 10.2514/6.1972-937

AstroRepresentations.convert6_coe_to_equi — Function

convert6_coe_to_equi(coe::AbstractVector{<:Number}, [args]...)

Convert classical orbital elements state vector to Equinoctial keplerian elements.

Inputs

coe – Keplerian elements – L, rad

Output

Equinoctial representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.

AstroRepresentations.convert6_equi_to_coe — Function

convert6_equi_to_coe(coe::AbstractVector{<:Number}, [args]...)

Convert Equinoctial keplerian elements to classical orbital elements state vector .

Inputs

equi – Equinoctial elements – L, rad

Output

Keplerian representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.

Spherical State Representation (sphe)

AstroRepresentations.convert6_cart_to_sphe — Function

convert6_cart_to_sphe(sv::AbstractVector{<:Number}, [args]...)

Convert cartesian state into spherical (ra-dec) representation.

AstroRepresentations.convert6_sphe_to_cart — Function

convert6_sphe_to_cart(sv::AbstractVector{<:Number}, [args]...)

Convert spherical (ra-dec) state into cartesian representation.

AstroRepresentations.convert3_cart_to_sphe — Function

convert3_cart_to_sphe(sv::AbstractVector{<:Number}, [args]...)

Convert cartesian position into spherical (ra-dec).

AstroRepresentations.convert3_sphe_to_cart — Function

convert6_sphe_to_cart(sv::AbstractVector{<:Number}, [args]...)

Convert spherical (ra-dec) position into cartesian.