SMDGraphs.jl API

SMDGraphs.AbstractGraphNode — Type

AbstractGraphNode

Abstract type for all graph nodes types.

SMDGraphs.MappedNodeGraph — Type

MappedNodeGraph{N, G}

Create a graph with mapped nodes.

Fields

graph – Graph

mid – Mapped id to nodes dictionary

nodes – Mapped nodes

edges – List of the edges between the nodes

paths – List of the available paths in the graph

Constructors

MappedNodeGraph{N}(g::G) where {G <: AbstractGraph, N <: AbstractGraphNode}

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(g::MappedNodeGraph, from::Int, to::Int, [cost])

Add an edge between from and to to g. Optionally assign a cost to the edge.

Graphs.SimpleGraphs.add_vertex! — Method

add_vertex!(g, node)

Add node to the graph g.

Graphs.has_path — Method

has_path(g, from, to)

Return true if there is a path between from and to in the graph g.

Graphs.has_vertex — Method

has_vertex(g, node)

Return true if node is contained in the graph g.

SMDGraphs.MappedDiGraph — Method

MappedDiGraph(::Type{N}) where {N}

Construct a directed MappedNodeGraph from node type N.

SMDGraphs.MappedGraph — Method

MappedGraph(::Type{N}) where {N}

Construct a MappedNodeGraph from node type N.

SMDGraphs.compute_paths! — Method

compute_paths(g::MappedNodeGraph)

Compute all possible paths in the graph.

SMDGraphs.get_edgecosts — Method

get_edgecosts(g::MappedNodeGraph, from::Int, to::Int)

Get all costs assigned to the edges between from and to. Returns an empty array if either from or to are not a part of g or if there is no path between them.

SMDGraphs.get_mappedid — Method

get_mappedid(g::MappedNodeGraph, node::Int)

Get the mappedid associated with a node.

SMDGraphs.get_mappednode — Method

get_mappednode(g::MappedNodeGraph, mid::Int)

Get the node associated to the given mapped id.

SMDGraphs.get_node — Method

get_node(g::MappedNodeGraph, node::Int)

Get the node associated with a node index.

SMDGraphs.get_node_id — Method

get_node_id(b::AbstractGraphNode)

Get the mapped-id of an AbstractGraphNode.

Warning

This method is abstract! A concrete implementation for each concrete node shall be defined.

SMDGraphs.get_path — Method

get_path(g::MappedNodeGraph, from::Int, to::Int)

Get the nodes on the path between and including from and to. Returns an empty array if either from or to are not a part of g or if there is no path between them.