Math

JSMDUtils.Math.InterpAkima — Type

InterpAkima{T, N} <: AbstractInterpolationMethod

Type storing an Akima spline nodes and coefficient. T is the interpolation data type and N is the spline dimension (i.e., the number of interpolated functions).

Fields

n – Number of node points.

xn – Interpolated node points.

yn – Node points function values

c – Akima polynomial coefficients.

JSMDUtils.Math.InterpAkima — Method

InterpAkima(x::AbstractVector, y::AbstractArray)

Construct an Akima spline interpolant from a set of data points x and their values y. Multi-dimensional splines can be constructed passing y as a subtype of AbstractMatrix such that each row contains a different set of values to be interpolated and the number of columns equals the number of data points.

References

Akima, H. (1970), A New Method of Interpolation and Smooth Curve Fitting Based on Local Procedures, Journal of the ACM, DOI:

JSMDUtils.Math.InterpCubicSplines — Type

InterpCubicSplines(x::AbstractVector, y::AbstractArray, type::Symbol=:Natural)

Construct a cubic spline interpolant from a set of data points x and their values y. Multi-dimensional splines can be constructed by passing y as a subtype of AbstractMatrix, such that each row contains a different set of values to be interpolated and the number of columns equals the number of data points.

Different boundary conditions can be applied based on the specified type:

:Natural: the second derivative of the first and the last polynomial are equal to zero at the boundary points.

:NotAKnot: the third derivatives of the first and last two polynomials are equal in the points where they meet each other.

:Periodic: the first and second derivatives at the initial and final points are equal.

:Quadratic: the first and the last polynomial are quadratic.

JSMDUtils.Math.InterpCubicSplines — Type

InterpCubicSplines{T, N} <: AbstractInterpolationMethod

Type storing a cubic spline nodes and coefficients. T is the spline data type and N is the spline dimension (i.e., the number of interpolated functions).

Fields

n – Number of node points.

xn – Interpolated node points.

yn – Node points function values.

c – Spline polynomials coefficients.

type – Boundary conditions type.

JSMDInterfaces.Math.interpolate — Method

interpolate(ak::InterpAkima, x::Number, flat::Bool=true)

Interpolate the Akima spline ak at point x. If the spline has a single dimension (e.g., InterpAkima{T, 1}), a scalar value is returned. Otherwise, an SVector is computed.

If x is outside the boundary range of sp a flat extrapolation is used by default. If the flat argument is false, the first and last polynomials will be used to compute all the outside values.

JSMDInterfaces.Math.interpolate — Method

interpolate(sp::InterpCubicSplines, x::Number, flat::Bool=true)

Interpolate the cubic spline sp at point x. If the spline has a single dimension (e.g., InterpCubicSpline{T, 1}), a scalar value is returned. Otherwise, an SVector is computed.

If x is outside the boundary range of sp a flat extrapolation is used by default. If the flat argument is false, the first and last polynomials will be used to compute all the outside values.

JSMDUtils.Math.D² — Method

D²(f, x::Real)

Return d²f/dx² evaluated at x using ForwardDiff, assuming f is called as f(x).

This method assumes that isa(f(x), Union{Real, AbstractArray}).

JSMDUtils.Math.D³ — Method

D³(f, x::Real)

Return d³f/dx³ evaluated at x using ForwardDiff, assuming f is called as f(x).

This method assumes that isa(f(x), Union{Real, AbstractArray}).

JSMDUtils.Math.D¹ — Method

D¹(f, x::Real)

Return df/dx evaluated at x using ForwardDiff, assuming f is called as f(x).

This method assumes that isa(f(x), Union{Real, AbstractArray}).

JSMDUtils.Math.skew — Function

skew(v::AbstractArray{N}) where {N<:Number}

Compute a skew-symmetric matrix from the vector v.

FileUtils

JSMDUtils.FileUtils.JSON — Type

JSMDUtils.FileUtils.JSON <: JSMDInterfaces.FilesIO.AbstractFile

A type representing JSMDUtils.FileUtils.JSON files.

JSMDUtils.FileUtils.TXT — Type

JSMDUtils.FileUtils.TXT <: JSMDInterfaces.FilesIO.AbstractFile

A type representing JSMDUtils.FileUtils.TXT files.

JSMDUtils.FileUtils.YAML — Type

JSMDUtils.FileUtils.YAML <: JSMDInterfaces.FilesIO.AbstractFile

A type representing JSMDUtils.FileUtils.YAML files.

JSMDInterfaces.FilesIO.load — Method

load(file::JSON{1})

Open a JSON file and parse its data in a dictionary.

JSMDInterfaces.FilesIO.load — Method

load(file::TXT{1})

Open a TEXT file and parse its data in a list of strings.

JSMDInterfaces.FilesIO.load — Method

load(file::YAML{1})

Open a YAML file and parse its data in a dictionary.

Miscellaneous

JSMDUtils.NullEphemerisProvider — Type

EmptyEphemerisProvider <: AbstractEphemerisProvider

Empty provider to initialise the frame system without loading ephemeris files.

JSMDUtils.format_camelcase — Method

format_camelcase(str::AbstractString)

Format str in CamelCase, such that the first letter of each word in the sentence is capitalized and spaces are removed.

JSMDUtils.format_snakecase — Method

format_snakecase(str::AbstractString)

Format str in SnakeCase, such that all the letters are in lower case and spaces are replaced with underscores.