API

SMDGraphs.MappedNodeGraph — Type

MappedNodeGraph{N, G} <: AbstractJSMDGraph{Int}

Create a graph with mapped nodes.

Fields

graph – Graph

mid – Mapped id to nodes dictionary

nodes – Mapped nodes

edges – List of the edges between the nodes

paths – List of the available paths in the graph

Constructors

MappedNodeGraph{N}(g::G) where {G <: AbstractGraph, N <: AbstractGraphNode}

SMDGraphs.MappedDiGraph — Method

MappedDiGraph(::Type{N}) where {N}

Construct a directed MappedNodeGraph from node type N.

SMDGraphs.MappedGraph — Method

MappedGraph(::Type{N}) where {N}

Construct a MappedNodeGraph from node type N.

SMDGraphs.add_edge_cost! — Method

add_edge_cost!(g::MappedNodeGraph{T, <:SimpleGraph}, fid::Int, tid::Int, cost::Int)

For a SimpleGraph type, register the edge cost between the nodes with mapped IDs fid and tid in both directions.

SMDGraphs.add_edge_cost! — Method

add_edge_cost!(g::MappedNodeGraph, fid::Int, tid::Int, cost::Int)

Register the cost between of the edge from the node with mapped ID fid to the node with mapped ID tid.

SMDGraphs.compute_paths! — Method

compute_paths(g::MappedNodeGraph)

Compute all possible paths in the graph.

SMDGraphs.get_edgecosts — Method

get_edgecosts(g::MappedNodeGraph, from::Int, to::Int)

Get all costs assigned to the edges between from and to. Returns an empty array if either from or to are not a part of g or if there is no path between them.

SMDGraphs.get_mappedid — Method

get_mappedid(g::MappedNodeGraph, node::Int)

Get the mappedid associated with a node. The mappedid is the internal ID that is assigned to the node within the graph.

SMDGraphs.get_mappednode — Method

get_mappednode(g::MappedNodeGraph, mid::Int)

Get the node associated to the given mapped id.

SMDGraphs.get_node — Method

get_node(g::MappedNodeGraph, node::Int)

Get the node associated with a node index.

SMDGraphs.get_node_id — Method

get_node_id(b::AbstractJSMDGraphNode)

Get the mapped-id of an AbstractJSMDGraphNode.

Warning

This method is abstract! A concrete implementation for each concrete node shall be defined.

SMDGraphs.get_outerid — Method

get_outerid(g::MappedNodeGraph, id::Int)

Return the id of the node associated to the mapped id id. The outer ID is the ID that is assigned to the node by the user.