Errors

JSMDInterfaces.Errors.AbstractGenericException — Type

AbstractGenericException

A supertype for all ecosystem related errors.

JSMDInterfaces.Errors.GenericError — Type

true

A type representing generic errors.

JSMDInterfaces.Errors.NotImplementedError — Type

true

A type representing not implemented errors.

JSMDInterfaces.Errors.@module_error — Macro

@module_error

Create a type representing an error associated to a specific module.

Ephemeris

JSMDInterfaces.Ephemeris.AbstractEphemerisProvider — Type

AbstractEphemerisProvider

Abstract type to represent ephemeris types.

JSMDInterfaces.Ephemeris.EphemAxesRecord — Type

EphemAxesRecord

Store an orientation record metadata.

Fields

target – NAIF ID of the target axes.

jd_start – Julian date of the first time.

jd_end – Julian date of the last time.

axes – NAIF ID of the reference axes.

JSMDInterfaces.Ephemeris.EphemPointRecord — Type

EphemPointRecord

Store a position record metadata.

Fields

target – NAIF ID of the target point.

center – NAIF ID of the center point.

jd_start – Julian date of the first time.

jd_end – Julian date of the last time.

axes – NAIF ID of the reference axes.

JSMDInterfaces.Ephemeris.EphemerisError — Type

true

A type representing ephemeris load or computations errors.

JSMDInterfaces.Ephemeris.ephem_compute! — Method

ephem_compute!(res, eph::E, jd0::Number, time::Number, target::Int, 
    center::Int, order::Int) where {E<:AbstractEphemerisProvider}

Abstract method to compute ephemeris.

Algorithm

Compute position and derivatives up to order of target with respect to center at epoch jd0+time.

Arguments

eph – ephemeris

jd0 – jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris.

time – jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris.

target – The body or reference point whose coordinates are required.

center – The origin of the coordinate system.

order – The order of derivatives from 0 (position) to 3 (position, velocity, acceleration and jerk).

JSMDInterfaces.Ephemeris.ephem_orient! — Method

ephem_orient!(res, eph::E, jd0::Number, time::Number, target::Int, 
    order::Int) where {E<:AbstractEphemerisProvider}

Abstract method to compute orientations.

Algorithm

Compute Euler angles and derivatives up to order for the orientation of target at epoch jd0 + time.

Arguments

eph: ephemeris

jd0: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris.

time: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris.

target: The body whose orientation is required.

order – The order of derivatives from 0 (position) to 3 (position, velocity, acceleration and jerk).

JSMDInterfaces.IO.load — Method

load(::Type{<:AbstractEphemerisProvider}, files::String...)

Load ephemeris files.

Math

JSMDInterfaces.Math.AbstractInterpolationMethod — Type

AbstractInterpolationMethod

Abstract type representing any interpolation method.

JSMDInterfaces.Math.interpolate — Method

interpolate(::AbstractInterpolationMethod, x)

Abstract interpolator call method.

IO

JSMDInterfaces.IO.filepath — Method

filepath(files::AbstractFile, idx::Int)

Return the path of the idx-th file within files.

JSMDInterfaces.IO.filepaths — Method

filepaths(files::AbstractFile)

Return the path of all the files loaded within file.

JSMDInterfaces.IO.load — Method

load(files::T) where T <: AbstractFile

Generic loader of different file/s formats.

JSMDInterfaces.IO.@filetype — Macro

filetype(ftype, suptype)

Create a type representing a specific file and the associated filepaths and filepath implementations.